+++ InstProfile/Posts&Stories/Models/Post.swift
@@ -9,6 +9,7 @@ import Foundation
 import UIKit
 
 // MARK: - Post model
+
 struct Post {
-    var postImage: UIImage
+    let postImage: UIImage
 }
