+++ InstProfile/Posts&Stories/Models/Story.swift
@@ -9,6 +9,7 @@ import Foundation
 import UIKit
 
 // MARK: - Story model
+
 struct Story {
     let storyImage: UIImage
     let storyText: String
