+++ InstProfile/Posts&Stories/Views/StoryCell.swift
@@ -6,16 +6,17 @@
 //
 
 import UIKit
-
+import SnapKit
 
 class StoryCell: UICollectionViewCell {
-    var storyImageView = UIImageView()
+    
+    // MARK: - constants
+    let storyImageView = UIImageView()
     let storyLabel = UILabel()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
         
-        // добавляю элементы на view
         addSubview(storyImageView)
         addSubview(storyLabel)
         
@@ -29,7 +30,7 @@ class StoryCell: UICollectionViewCell {
         fatalError("init(coder:) has not been implemented")
     }
     
-    // функция выполняет задачу обновления интерфейсных элементов на экране информацией из объекта Expense, переданного в качестве параметра
+    // функция выполняет задачу обновления интерфейсных элементов на экране информацией из объекта Story, переданного в качестве параметра
     func set(storiesArray: Story) {
         storyImageView.image = storiesArray.storyImage
         storyLabel.text      = storiesArray.storyText
@@ -39,9 +40,7 @@ class StoryCell: UICollectionViewCell {
     
     // настройка изображения
     func configureImageView() {
-        // устанавливает радиус скругления углов
         storyImageView.layer.cornerRadius = 30
-        // содержимое контейнера будет отображаться только в пределах его рамок, отсекая все, что находится за его пределами
         storyImageView.clipsToBounds = true
         storyImageView.contentMode = .scaleAspectFill
         storyImageView.translatesAutoresizingMaskIntoConstraints = false
@@ -52,31 +51,26 @@ class StoryCell: UICollectionViewCell {
     func congigureLabel() {
         // динамическое расширение линий
         storyLabel.numberOfLines = 0
-        // цвет текста
         storyLabel.textColor = .white
-        // шрифт
         storyLabel.font = UIFont.systemFont(ofSize: 15, weight: .medium)
         // уменьшает размер текста по пространству
         storyLabel.adjustsFontSizeToFitWidth = true
-        // отключение автоматических констрейтов
         storyLabel.translatesAutoresizingMaskIntoConstraints = false
     }
     
-    // MARK: - Constraits
+    // MARK: - Constraits using SnapKit
     
     func setConstraits() {
-        // Настройка constraints для imageView
-        NSLayoutConstraint.activate([
-            storyImageView.topAnchor.constraint(equalTo: topAnchor),
-            storyImageView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            storyImageView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            storyImageView.heightAnchor.constraint(equalToConstant: 60),
-            storyImageView.widthAnchor.constraint(equalToConstant: 60),
-            
-            storyLabel.topAnchor.constraint(equalTo: storyImageView.bottomAnchor, constant: 10),
-            storyLabel.bottomAnchor.constraint(equalTo: bottomAnchor),
-            storyLabel.centerXAnchor.constraint(equalTo: storyImageView.centerXAnchor)
-        ])
+        storyImageView.snp.makeConstraints { make in
+            make.top.trailing.leading.equalToSuperview()
+            make.height.width.equalTo(60)
+        }
+        
+        storyLabel.snp.makeConstraints { make in
+            make.top.equalTo(storyImageView.snp.bottom).offset(10)
+            make.bottom.equalToSuperview()
+            make.centerX.equalTo(storyImageView.snp.centerX)
+        }
     }
 }
